@@ -7,14 +7,14 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@faker-js/faker": "^7.2.0",
-    "@vitejs/plugin-react": "^1.0.7",
-    "msw": "^0.36.3",
-    "vite": "^2.7.2"
+    "@faker-js/faker": "^7.6.0",
+    "@vitejs/plugin-react": "^3.0.1",
+    "msw": "^1.0.0",
+    "vite": "^4.0.4"
   },
   "msw": {
     "workerDirectory": "public"
