@@ -1,16 +1,14 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.css';
+
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+
 import App from './App';
 import { startWorker } from './mock';
 
 function mountApp() {
-  ReactDOM.render(
-    <React.StrictMode>
-      <App />
-    </React.StrictMode>,
-    document.getElementById('root')
-  );
+  const root = createRoot(document.getElementById('root'));
+  root.render(<App />);
 }
 
 if (process.env.NODE_ENV === 'development') {
