@@ -1,17 +1,31 @@
 import { CliOptions } from './types';
+import { OperationCollection, transformToHandlerCode, transformToResObject } from './transform';
 
-export const mockTemplate = (
-  handlersCode: string,
-  baseURL: string,
-  options?: CliOptions
-) => `/**
+const getSetupCode = (isNode?: boolean) => {
+  if (isNode) {
+    return [`const server = setupServer(...handlers);`, `server.listen();`].join('\n');
+  }
+
+  return [`const worker = setupWorker(...handlers);`, `worker.start();`].join('\n');
+};
+
+const getImportsCode = (options?: CliOptions) => {
+  const imports = [`import { setupWorker, rest } from 'msw';`, `import { faker } from '@faker-js/faker';`];
+
+  if (options?.node) {
+    imports.push(`import { setupServer } from 'msw/node'`);
+  }
+
+  return imports.join('\n');
+};
+
+export const mockTemplate = (operationCollection: OperationCollection, baseURL: string, options?: CliOptions) => `/**
 * This file is AUTO GENERATED by [msw-auto-mock](https://github.com/zoubingwu/msw-auto-mock)
 * Feel free to commit/edit it as you need.
 */
 /* eslint-disable */
 /* tslint:disable */
-import { setupWorker, rest } from 'msw';
-import { faker } from '@faker-js/faker';
+${getImportsCode(options)}
 
 faker.seed(1);
 
@@ -27,18 +41,13 @@ const next = () => {
 }
 
 export const handlers = [
-  ${handlersCode}
+  ${transformToHandlerCode(operationCollection)}
 ];
+
+${transformToResObject(operationCollection)}
 
 // This configures a Service Worker with the given request handlers.
 export const startWorker = () => {
-  if (typeof window === 'undefined') {
-    const { setupServer } = require('msw/node');
-    const server = setupServer(...handlers);
-    server.listen();
-  } else {
-    const worker = setupWorker(...handlers);
-    worker.start();
-  }
+  ${getSetupCode(options?.node)}
 }
 `;
