@@ -1,26 +1,55 @@
 import { OpenAPIV3 } from 'openapi-types';
+import merge from 'lodash/merge';
+import camelCase from 'lodash/camelCase';
 
-export type OperationCollection = {
+export interface ResponseMap {
+  code: string;
+  id: string;
+  responses?: Record<string, OpenAPIV3.SchemaObject>;
+}
+
+export interface Operation {
   verb: string;
   path: string;
-  responseMap: {
-    code: string;
-    responses?: Record<string, OpenAPIV3.SchemaObject>;
-  }[];
-}[];
+  response: ResponseMap[];
+}
 
-export function transformToHandlerCode(
-  operationCollection: OperationCollection
-): string {
+export type OperationCollection = Operation[];
+
+export function getResIdentifierName(res: ResponseMap) {
+  if (!res.id) {
+    return '';
+  }
+  return camelCase(`get ${res.id}${res.code}Response`);
+}
+
+export function transformToResObject(operationCollection: OperationCollection): string {
+  return operationCollection
+    .map(op =>
+      op.response
+        .map(r => {
+          const name = getResIdentifierName(r);
+          if (!name) {
+            return '';
+          }
+          return `export function ${getResIdentifierName(r)}() { return ${transformJSONSchemaToFakerCode(
+            r.responses?.['application/json']
+          )} };\n`;
+        })
+        .join('\n')
+    )
+    .join('\n');
+}
+
+export function transformToHandlerCode(operationCollection: OperationCollection): string {
   return operationCollection
     .map(op => {
-      return `rest.${op.verb}(\`\${baseURL}${op.path}\`, (req, res, ctx) => {
-        const resultArray = [${op.responseMap.map(response => {
-          return `[ctx.status(${parseInt(
-            response?.code!
-          )}), ctx.json(${transformJSONSchemaToFakerCode(
-            response?.responses?.['application/json']
-          )})]`;
+      return `rest.${op.verb}(\`\${baseURL}${op.path}\`, (_, res, ctx) => {
+        const resultArray = [${op.response.map(response => {
+          const identifier = getResIdentifierName(response);
+          return parseInt(response?.code!) === 204
+            ? `[ctx.status(${parseInt(response?.code!)})]`
+            : `[ctx.status(${parseInt(response?.code!)}), ctx.json(${identifier ? `${identifier}()` : 'null'})]`;
         })}];
 
           return res(...resultArray[next() % resultArray.length])
@@ -30,10 +59,7 @@ export function transformToHandlerCode(
     .trimEnd();
 }
 
-function transformJSONSchemaToFakerCode(
-  jsonSchema?: OpenAPIV3.SchemaObject,
-  key?: string
-): string {
+function transformJSONSchemaToFakerCode(jsonSchema?: OpenAPIV3.SchemaObject, key?: string): string {
   if (!jsonSchema) {
     return 'null';
   }
@@ -52,20 +78,31 @@ function transformJSONSchemaToFakerCode(
     return `faker.helpers.arrayElement(${JSON.stringify(jsonSchema.enum)})`;
   }
 
+  if (jsonSchema.allOf) {
+    const schemas = jsonSchema.allOf as OpenAPIV3.SchemaObject[];
+    return transformJSONSchemaToFakerCode(merge({}, ...schemas));
+  }
+
+  if (jsonSchema.oneOf) {
+    const schemas = jsonSchema.oneOf as OpenAPIV3.SchemaObject[];
+    return `faker.helpers.arrayElement([${schemas.map(i => transformJSONSchemaToFakerCode(i))}])`;
+  }
+
+  if (jsonSchema.anyOf) {
+    const schemas = jsonSchema.anyOf as OpenAPIV3.SchemaObject[];
+    return `faker.helpers.arrayElement([${schemas.map(i => transformJSONSchemaToFakerCode(i))}])`;
+  }
+
   switch (jsonSchema.type) {
     case 'string':
       return transformStringBasedOnFormat(jsonSchema.format, key);
     case 'number':
-      return `faker.datatype.number()`;
     case 'integer':
-      return `faker.datatype.number()`;
+      return `faker.datatype.number({ min: ${jsonSchema.minimum}, max: ${jsonSchema.maximum} })`;
     case 'boolean':
       return `faker.datatype.boolean()`;
     case 'object':
-      if (
-        !jsonSchema.properties &&
-        typeof jsonSchema.additionalProperties === 'object'
-      ) {
+      if (!jsonSchema.properties && typeof jsonSchema.additionalProperties === 'object') {
         return `[...new Array(5).keys()].map(_ => ({ [faker.lorem.word()]: ${transformJSONSchemaToFakerCode(
           jsonSchema.additionalProperties as OpenAPIV3.SchemaObject
         )} })).reduce((acc, next) => Object.assign(acc, next), {})`;
@@ -74,17 +111,14 @@ function transformJSONSchemaToFakerCode(
       return `{
         ${Object.entries(jsonSchema.properties ?? {})
           .map(([k, v]) => {
-            return `${JSON.stringify(k)}: ${transformJSONSchemaToFakerCode(
-              v as OpenAPIV3.SchemaObject,
-              k
-            )}`;
+            return `${JSON.stringify(k)}: ${transformJSONSchemaToFakerCode(v as OpenAPIV3.SchemaObject, k)}`;
           })
           .join(',\n')}
     }`;
     case 'array':
-      return `[...(new Array(faker.datatype.number({ max: MAX_ARRAY_LENGTH }))).keys()].map(_ => (${transformJSONSchemaToFakerCode(
-        jsonSchema.items as OpenAPIV3.SchemaObject
-      )}))`;
+      return `[...(new Array(faker.datatype.number({ min: ${jsonSchema.minLength ?? 1}, max: ${
+        jsonSchema.maxLength ?? 'MAX_ARRAY_LENGTH'
+      } }))).keys()].map(_ => (${transformJSONSchemaToFakerCode(jsonSchema.items as OpenAPIV3.SchemaObject)}))`;
     default:
       return 'null';
   }
@@ -94,17 +128,11 @@ function transformJSONSchemaToFakerCode(
  * See https://json-schema.org/understanding-json-schema/reference/string.html#built-in-formats
  */
 function transformStringBasedOnFormat(format?: string, key?: string) {
-  if (
-    ['date-time', 'date', 'time'].includes(format ?? '') ||
-    key?.toLowerCase().endsWith('_at')
-  ) {
+  if (['date-time', 'date', 'time'].includes(format ?? '') || key?.toLowerCase().endsWith('_at')) {
     return `faker.date.past()`;
   } else if (format === 'uuid') {
     return `faker.datatype.uuid()`;
-  } else if (
-    ['idn-email', 'email'].includes(format ?? '') ||
-    key?.toLowerCase().endsWith('email')
-  ) {
+  } else if (['idn-email', 'email'].includes(format ?? '') || key?.toLowerCase().endsWith('email')) {
     return `faker.internet.email()`;
   } else if (['hostname', 'idn-hostname'].includes(format ?? '')) {
     return `faker.internet.domainName()`;
@@ -113,9 +141,7 @@ function transformStringBasedOnFormat(format?: string, key?: string) {
   } else if (format === 'ipv6') {
     return `faker.internet.ipv6()`;
   } else if (
-    ['uri', 'uri-reference', 'iri', 'iri-reference', 'uri-template'].includes(
-      format ?? ''
-    ) ||
+    ['uri', 'uri-reference', 'iri', 'iri-reference', 'uri-template'].includes(format ?? '') ||
     key?.toLowerCase().endsWith('url')
   ) {
     return `faker.internet.url()`;
